+++ src/plugins/index.ts
@@ -1,8 +1,10 @@
-import { IPlugin } from '@/iplugin.js';
 import * as amazon from './amazon.js';
 import * as wikipedia from './wikipedia.js';
+import * as branchIoDeeplinks from './branchio-deeplinks.js';
+import { SummalyPlugin } from '@/iplugin.js';
 
-export const plugins: IPlugin[] = [
-    amazon,
-    wikipedia,
+export const plugins: SummalyPlugin[] = [
+	amazon,
+	wikipedia,
+	branchIoDeeplinks,
 ];