+++ src/plugins/wikipedia.ts
@@ -1,4 +1,5 @@
 import { URL } from 'node:url';
+import debug from 'debug';
 import { get } from '../utils/got.js';
 import summary from '../summary.js';
 import clip from './../utils/clip.js';
@@ -13,11 +14,16 @@ export async function summarize(url: URL): Promise<summary> {
 	const title = url.pathname ? url.pathname.split('/')[2] : null;
 	const endpoint = `https://${lang}.wikipedia.org/w/api.php?format=json&action=query&prop=extracts&exintro=&explaintext=&titles=${title}`;
 
+	log(`lang is ${lang}`);
+	log(`title is ${title}`);
+	log(`endpoint is ${endpoint}`);
+
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	let body = await get(endpoint) as any;
 	body = JSON.parse(body);
 
 	if (!('query' in body) || !('pages' in body.query)) {
-		throw 'fetch failed';
+		throw new Error('fetch failed');
 	}
 
 	const info = body.query.pages[Object.keys(body.query.pages)[0]];